TST: Further shaping of testing for this extensions
